refactor(ProductRepository): Product Repository interface 생성
